+++ azure/storage/blob/blockblobservice.py
@@ -790,6 +790,41 @@ class BlockBlobService(BaseBlobService):
                 if_none_match=if_none_match,
                 timeout=timeout)
 
+    def set_standard_blob_tier(
+        self, container_name, blob_name, standard_blob_tier, timeout=None):
+        '''
+        Sets the block blob tiers on the blob. This API is only supported for block blobs on standard storage accounts.
+
+        :param str container_name:
+            Name of existing container.
+        :param str blob_name:
+            Name of blob to update.
+        :param StandardBlobTier standard_blob_tier:
+            A standard blob tier value to set the blob to. For this version of the library,
+            this is only applicable to block blobs on standard storage accounts.
+        :param int timeout:
+            The timeout parameter is expressed in seconds. This method may make
+            multiple calls to the Azure service and the timeout will apply to
+            each call individually.
+        '''
+        _validate_not_none('container_name', container_name)
+        _validate_not_none('blob_name', blob_name)
+        _validate_not_none('standard_blob_tier', standard_blob_tier)
+
+        request = HTTPRequest()
+        request.method = 'PUT'
+        request.host_locations = self._get_host_locations()
+        request.path = _get_path(container_name, blob_name)
+        request.query = {
+            'comp': 'tier',
+            'timeout': _int_to_str(timeout),
+        }
+        request.headers = {
+            'x-ms-access-tier': _to_str(standard_blob_tier)
+        }
+
+        self._perform_request(request)
+
     #-----Helper methods------------------------------------
     def _put_blob(self, container_name, blob_name, blob, content_settings=None,
                   metadata=None, validate_content=False, lease_id=None, if_modified_since=None,